LayoutTests/http/tests/performance/performance-resource-timing-cached-entries-expected.txt

CONSOLE MESSAGE: line 11: current entries: 0
CONSOLE MESSAGE: line 20: entry: http://127.0.0.1:8000/resources/js-test-pre.js
CONSOLE MESSAGE: line 20: entry: http://127.0.0.1:8000/resources/square100.png
CONSOLE MESSAGE: line 20: entry: http://127.0.0.1:8000/resources/square100.png
PASS foundResource is 2
